More work

+++ scripts/shibboleth_install.sh
@@ -1,9 +1,14 @@
-# Add Shibboleth repository (Java Service Provider)
-cd /etc/yum.repos.d/
-wget http://download.opensuse.org/repositories/security://shibboleth/RHEL_6/security:shibboleth.repo
-cd -
+# Add Shibboleth repository (Java SAML Service Provider)
 
-# install non-CentOS packages
-yum -y install shibboleth
+yum -q list shibboleth > /dev/null
+if [ $? -ne 0 ]
+then
+    cd /etc/yum.repos.d/
+    wget http://download.opensuse.org/repositories/security://shibboleth/RHEL_6/security:shibboleth.repo
+    cd -
 
-rm -vf /etc/yum.repos.d/security\:shibboleth.repo
+    # install non-CentOS packages
+    yum -y -q install shibboleth
+
+    rm -vf /etc/yum.repos.d/security\:shibboleth.repo
+fi
